fix: проект в +- в нормальном состоянии

+++ src/main/java/model/ExchangeRateResponse.java
@@ -3,11 +3,13 @@ package model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ExchangeRateResponse {
     private Long id;
     private Currency baseCurrency;
     private Currency targetCurrency;
-    private Double rate;
+    private BigDecimal rate;
 }
